--- core/liteblog.py
@@ -1,318 +1,0 @@
-from .lite import DBLite, one_factory
-from .writer import MDWriter
-import sqlite3
-from bunch import Bunch
-from functools import lru_cache
-from textwrap import dedent
-from .schemaspy import SchemasPy
-import re
-
-re_blank = re.compile(r'^\s*$\n', re.MULTILINE)
-
-def get_dom(url):
-    if url is None:
-        return None
-    slp = url.split("://", 1)
-    if len(slp)!=2 and slp[0].lower() not in ("http", "https"):
-        return None
-    dom = slp[1]
-    dom = dom.split("/", 1)[0]
-    return dom
-
-def build_tr(md_row, space=" "):
-    md_row = md_row.strip()
-    md_row = (i.strip() for i in md_row.split("\n"))
-    md_row = (space+"|"+space).join(md_row)
-    md_row = "|"+space+md_row+space+"|"
-    return md_row
-
-def write(f, sql, *args, end=None):
-    if end is None:
-        end = ";\n\n"
-    if len(args) > 0:
-        sql = sql.format(*args)
-    sql = re_blank.sub("", sql)
-    sql = dedent(sql).strip()
-    f.write(sql+end)
-
-class BlogDBLite(DBLite):
-    def __init__(self, *args, total=None, **kargv):
-        super().__init__(*args, **kargv)
-        self.con.create_function("get_dom", 1, get_dom)
-        self.id_blogs={}
-        self.total = total
-        self.count = 0
-
-    def get_blog_id(self, blog):
-        if blog not in self.id_blogs:
-            self.id_blogs[blog] = self.one("select ID from blogs where url = '" + blog + "'")
-        if blog not in self.id_blogs:
-            raise Exception("No se ha encontrad el ID del blog "+blog)
-        return self.id_blogs[blog]
-
-    def insert(self, table, **kargv):
-        if table == "blogs":
-            if kargv.get("ID") is None:
-                id = self.one("select max(ID) from blogs")
-                kargv["ID"] = (id or 0)+1
-            if "url" in kargv:
-                self.id_blogs[kargv["url"]] = kargv["ID"]
-        else:
-            for k in ("blog", "in_blog"):
-                if k in kargv and isinstance(kargv[k], str):
-                    kargv[k] = self.get_blog_id(kargv[k])
-        super().insert(table, **kargv)
-        if self.total is not None:
-            self.count = self.count + 1
-            print("Creando sqlite {0:.0f}%".format(
-                self.count*100/self.total), end="\r")
-
-    @property
-    @lru_cache(maxsize=None)
-    def _ok_objects(self):
-        if "objects" not in self.tables:
-            return False
-        try:
-            list(self.select("select * from objects limit 1"))
-        except sqlite3.OperationalError:
-            return False
-        return True
-
-    @property
-    @lru_cache(maxsize=None)
-    def _sql_get_links(self):
-        if "urls" in self.tables:
-            return "select distinct url from urls where url is not null and url!='#'"
-        col_object_comment = "object"
-        if col_object_comment not in self.tables["comments"]:
-            col_object_comment = "post"
-        if not self._ok_objects:
-            if "url" not in self.tables["media"] and "url" not in self.tables["posts"]:
-                return None
-            return '''
-                select
-                    distinct url
-                from (
-                    select url, type, blog, ID from posts
-                    union
-                    select url, type, blog, ID from media
-                ) t where
-                    url is not null and
-                    url!='#' and (
-                        type!='pmedia' or
-                        (blog, ID) in (select blog, {0} from comments)
-                    )
-                order by
-                    blog, ID
-            '''.format(col_object_comment)
-        return '''
-            select
-                distinct url
-            from
-                objects
-            where
-                url is not null and
-                url!='#' and (
-                    type!='pmedia' or
-                    (blog, ID) in (select blog, {} from comments)
-                )
-            order by
-                blog, ID
-        '''.format(col_object_comment)
-
-    def print_links(self, file):
-        if self._sql_get_links is None:
-            return
-        with open(file, "w") as f:
-            for l in self.select(self._sql_get_links, row_factory=one_factory):
-                f.write(l+"\n")
-
-    def get_info(self, blog=None):
-        where = ""
-        if blog is not None:
-            where = " where blog = "+str(blog)
-        fch = []
-        for t, c in self.find_cols("date"):
-            if blog is not None and "blog" not in self.tables[t]:
-                continue
-            fch.append("select substr({0}, 1, 10) d from {1}".format(c, t)+where)
-        fch = '''
-        select
-            min(d) ini,
-            max(d) fin
-        from (
-            {0}
-        ) t
-        where
-            d>='2011-05-17' and
-            d<=date('now')
-        '''.format(" union \n".join(fch))
-        ini, fin = self.one(fch)
-        r = Bunch(
-            ini=ini,
-            fin=fin,
-            counts={}
-        )
-        tables = ('blogs', 'posts', 'media', 'comments', 'tags')
-        for t in tables:
-            if t not in self.tables:
-                continue
-            if blog is not None and "blog" not in self.tables[t]:
-                continue
-            r.counts[t]=self.one("select count(*) from "+t+where)
-        return r
-
-    def print_info(self, file, table_link=True):
-        md=MDWriter(file)
-        r = self.get_info()
-        md.write("Datos desde el {ini} al {fin}:".format(**dict(r)), end="\n\n")
-        for t, c in r.counts.items():
-            md.write("* {rows} registros en `{table}`".format(table=t, rows=c))
-        if "url" in self.tables["blogs"]:
-            def _sort_blogs(x):
-                id, url = x
-                r = url.split("/", 1)
-                r[0]=tuple(reversed(r[0].split(".")))
-                if len(r)==1:
-                    r.append("")
-                r.append(id)
-                return tuple(r)
-
-            if self._sql_get_links is not None:
-                urls = self.one("select count(*) from ("+self._sql_get_links+")")
-                md.write("")
-                md.write("Lo que supone {urls} urls.".format(urls=urls))
-            blogs = self.select("select id, url from blogs")
-            blogs = sorted(blogs, key=_sort_blogs)
-            max_blog = max(len(x[1]) for x in blogs)
-            md.write("")
-            md.write(dedent('''
-                | BLOG | post/page | Comentarios | Último uso | 1º uso | Último comentario |
-                |:-----|----------:|------------:|-----------:|-------:|------------------:|
-            ''').strip())
-            for id, url in blogs:
-                info = self.get_info(id)
-                row = dict(info.counts)
-                row["ini"] = info.ini
-                row["fin"] = info.fin
-                row["site"] = url
-                row["admin"] = "https://{}/wp-admin/".format(url)
-                row["ult_comment"] = self.one("select IFNULL(substr(date, 1, 10), '') from comments")
-                if table_link:
-                    md_row = build_tr('''
-                        [{site}](https://{site})
-                        [{posts}]({admin}edit.php?orderby=date&order=desc)
-                        [{comments}]({admin}edit-comments.php?comment_type=comment&orderby=comment_date&order=desc)
-                        {fin}
-                        {ini}
-                        {ult_comment}
-                    ''', space="")
-                else:
-                    md_row = build_tr('''
-                        {site:<%s}
-                        {posts:>6}
-                        {comments:>6}
-                        {fin}
-                        {ini}
-                        {ult_comment}
-                    ''' % max_blog)
-                md.write(md_row.format(**row).strip())
-            md.write("")
-            if table_link:
-                md.write(dedent('''
-                    Para reordenar la tabla puede usar las extensiones
-                    [`Tampermonkey`](https://chrome.google.com/webstore/detail/tampermonkey/dhdgffkkebhmkfjojejmpbldmpobfkfo?hl=es)
-                    o [`Greasemonkey`](https://addons.mozilla.org/es/firefox/addon/greasemonkey/)
-                    con [`Github Sort Content`](https://greasyfork.org/en/scripts/21373-github-sort-content)
-                '''))
-        md.close
-
-    def minimize(self, file):
-        re_rem = re.compile(r"^\s*_.*$", re.MULTILINE)
-        with open(file, "w") as f:
-            write(f, "BEGIN TRANSACTION")
-            write(f, "DELETE from tags where (blog, post) in (select blog, id from posts where url is null)")
-            sql = "SELECT type, name FROM sqlite_master WHERE type in ('view', 'table') and name like '^_%' ESCAPE '^'"
-            for r in self.select(sql):
-                write(f, "DROP {0} {1}", *r, end=";\n")
-            write(f, "", end="\n")
-            tables = [i for i in self.tables.items() if not i[0].startswith("_")]
-            flag = False
-            for t, cs in tables:
-                if "_delete" in cs:
-                    write(f, "DELETE from {0} where _delete=1", t, end=";\n")
-                    flag = True
-            if flag:
-                f.write("\n")
-            for t, cs in tables:
-                ok = []
-                ko = []
-                for c in cs:
-                    if c.startswith("_"):
-                        ko.append(c)
-                    else:
-                        ok.append(c)
-                if len(ko) > 0:
-                    write(f, "ALTER TABLE {0} RENAME TO temp_{0}", t)
-                    sql = self.get_sql_table(t)
-                    sql = re_rem.sub("", sql)
-                    write(f, sql)
-                    ok_select = list(ok)
-                    if t == "posts" and "content" in ok_select:
-                        ok_select[ok_select.index("content")] = "_content"
-                    end = None
-                    if "url" in ok_select:
-                        end = "\nwhere url is not null;\n\n"
-                    write(f, '''
-                        INSERT INTO {0}
-                            ({1})
-                        SELECT
-                            {2}
-                        FROM
-                            temp_{0}
-                    ''', t, ", ".join(ok), ", ".join(ok_select), end=end)
-                    write(f, "DROP TABLE temp_{0}", t)
-            write(f, "COMMIT", end=";")
-
-        self.execute(file)
-
-    def save_diagram(self, file, *args, **kargv):
-        sch = SchemasPy()
-        sch.save_diagram(self.file, file, "-noviews", *args, **kargv)
-
-if __name__ == "__main__":
-    import sys
-    from glob import glob
-    from .util import unzip
-    from shutil import copyfile, move
-    from os.path import dirname, isfile, basename, isdir
-    from os import makedirs, remove
-    for db in sys.argv[1:]:
-        pre = db.split("/")
-        pre = pre[-1]
-        pre = pre.split(".")
-        pre = pre[0]
-        pre_dt = pre[:10].replace("-", ".")
-        print(db,"->", end=" ")
-        db = unzip(db, get="*.db")
-        print(db)
-        tg = dirname(db)+"/{pre_dt}_wp.db".format(pre_dt=pre_dt)
-        if not isfile(tg):
-            print(db, "->", tg)
-            move(db, tg)
-        db.close(vacuum=True)
-        target = "data/archive/{pre}/wp.7z".format(pre=pre)
-        db.zip(target, file=tg)
-        print(tg,"->", target)
-        remove(tg)
-    for zip in sorted(glob("data/archive/*/*.7z")):
-        print(zip,"->", end=" ")
-        target = dirname(zip) + "/"
-        tg = unzip(zip, get="*.db")
-        print(basename(tg))
-        db = BlogDBLite(tg, readonly=True)
-        db.print_links(target+"links.txt")
-        db.print_info(target+"README.md", table_link=False)
-        db.save_diagram(target+"diagram.png")
-        db.close()
-        remove(tg)
