+++ data/dict.txt
@@ -49,7 +49,7 @@ takethesquare.net
 192 https://15oct.takethesquare.net/files/2011/09/spread2.png
 marchestobrussels.takethesquare.net
  122 https://marchestobrussels.takethesquare.net/files/2011/08/boton_facebook.png
- 123 https://marchestobrussels.takethesquare.net/files/2011/08/boton_flickr.png
+ 123 http://marchestobrussels.takethesquare.net/files/2011/08/boton_flickr.png
  124 https://marchestobrussels.takethesquare.net/files/2011/08/boton_twitter.png
  125 https://marchestobrussels.takethesquare.net/files/2011/08/rss-e1313605950984.png
 2279 https://marchestobrussels.takethesquare.net/files/2011/10/SIGNOS-ASAMBLEARIOS-copia.jpg
