+++ sql/clean.sql
@@ -1,32 +1,32 @@
 BEGIN TRANSACTION;
 
-DELETE from tags where (blog, post) in (select blog, id from posts where url is null);
+DELETE from tags where (site, post) in (select site, id from posts where url is null);
 
 DROP view _posts;
 DROP view _media;
 
-ALTER TABLE blogs RENAME TO temp_blogs;
+ALTER TABLE sites RENAME TO temp_sites;
 
-CREATE TABLE blogs (
+CREATE TABLE sites (
   ID INTEGER,
   url TEXT,
   PRIMARY KEY (ID)
 );
 
-INSERT INTO blogs
+INSERT INTO sites
     (ID, url)
 SELECT
     ID, url
 FROM
-    temp_blogs
+    temp_sites
 where url is not null;
 
-DROP TABLE temp_blogs;
+DROP TABLE temp_sites;
 
 ALTER TABLE posts RENAME TO temp_posts;
 
 CREATE TABLE posts (
-  blog INTEGER REFERENCES blogs(ID),
+  site INTEGER REFERENCES sites(ID),
   ID INTEGER,
   type TEXT,
   date TEXT,
@@ -35,13 +35,13 @@ CREATE TABLE posts (
   name TEXT,
   author TEXT,
   url TEXT,
-  PRIMARY KEY (blog, id)
+  PRIMARY KEY (site, id)
 );
 
 INSERT INTO posts
-    (blog, ID, type, date, content, title, name, author, url)
+    (site, ID, type, date, content, title, name, author, url)
 SELECT
-    blog, ID, type, date, _content, title, name, author, url
+    site, ID, type, date, _content, title, name, author, url
 FROM
     temp_posts
 where url is not null;
@@ -51,7 +51,7 @@ DROP TABLE temp_posts;
 ALTER TABLE media RENAME TO temp_media;
 
 CREATE TABLE media (
-  blog INTEGER REFERENCES blogs(ID),
+  site INTEGER REFERENCES sites(ID),
   ID INTEGER,
   type TEXT,
   date TEXT,
@@ -59,13 +59,13 @@ CREATE TABLE media (
   file TEXT,
   url TEXT,
   page TEXT,
-  PRIMARY KEY (blog, id)
+  PRIMARY KEY (site, id)
 );
 
 INSERT INTO media
-    (blog, ID, type, date, author, file, url, page)
+    (site, ID, type, date, author, file, url, page)
 SELECT
-    blog, ID, type, date, author, file, url, page
+    site, ID, type, date, author, file, url, page
 FROM
     temp_media
 where url is not null;
@@ -76,19 +76,19 @@ ALTER TABLE comments RENAME TO temp_comments;
 
 CREATE TABLE comments (
   ID INTEGER,
-  blog INTEGER REFERENCES blogs(id),
+  site INTEGER REFERENCES sites(id),
   object INTEGER,
   content TEXT,
   date TEXT,
   author TEXT,
   parent INTEGER,
-  PRIMARY KEY (ID, blog, object)
+  PRIMARY KEY (ID, site, object)
 );
 
 INSERT INTO comments
-    (ID, blog, object, content, date, author, parent)
+    (ID, site, object, content, date, author, parent)
 SELECT
-    ID, blog, object, content, date, author, parent
+    ID, site, object, content, date, author, parent
 FROM
     temp_comments;
